+++ packages/aws-appsync-nodejs/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.4](https://github.com/ttoss/ttoss/compare/@ttoss/aws-appsync-nodejs@1.7.3...@ttoss/aws-appsync-nodejs@1.7.4) (2023-12-12)
+
+### Bug Fixes
+
+- upload template when size is big ([#453](https://github.com/ttoss/ttoss/issues/453)) ([59371b7](https://github.com/ttoss/ttoss/commit/59371b7eb9befa006ffe316541f60dfc534dcbd4))
+
 ## [1.7.3](https://github.com/ttoss/ttoss/compare/@ttoss/aws-appsync-nodejs@1.7.2...@ttoss/aws-appsync-nodejs@1.7.3) (2023-12-01)
 
 **Note:** Version bump only for package @ttoss/aws-appsync-nodejs