+++ packages/config/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.31.4](https://github.com/ttoss/ttoss/compare/@ttoss/config@1.31.3...@ttoss/config@1.31.4) (2023-12-12)
+
+### Bug Fixes
+
+- upload template when size is big ([#453](https://github.com/ttoss/ttoss/issues/453)) ([59371b7](https://github.com/ttoss/ttoss/commit/59371b7eb9befa006ffe316541f60dfc534dcbd4))
+
 ## [1.31.3](https://github.com/ttoss/ttoss/compare/@ttoss/config@1.31.2...@ttoss/config@1.31.3) (2023-12-01)
 
 ### Bug Fixes