+++ packages/layouts/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.27](https://github.com/ttoss/ttoss/compare/@ttoss/layouts@0.1.26...@ttoss/layouts@0.1.27) (2023-12-12)
+
+### Bug Fixes
+
+- upload template when size is big ([#453](https://github.com/ttoss/ttoss/issues/453)) ([59371b7](https://github.com/ttoss/ttoss/commit/59371b7eb9befa006ffe316541f60dfc534dcbd4))
+
 ## [0.1.26](https://github.com/ttoss/ttoss/compare/@ttoss/layouts@0.1.25...@ttoss/layouts@0.1.26) (2023-12-05)
 
 **Note:** Version bump only for package @ttoss/layouts